requirements.txt
Django==4.2.5
django-cors-headers==4.2.0
django-filter==23.3
djangorestframework==3.14.0
drf-yasg==1.21.7
Pillow==10.0.1
PyYAML==6.0.1
requests==2.31.0
stripe==6.6.0